Corrigindo Fragments e arrumando API

+++ app/src/main/java/br/com/cotemig/projetofinal/models/Itens.kt
@@ -4,6 +4,6 @@ data class Itens (
     var id : Int = 0,
     var nome : String = "",
     var descricao : String = "",
-    var preco : Float = 0.0f,
+    var preco : Double = 0.0,
     var foto : String = ""
 )
